Open upgrade dialog from sidebar
