+++ apps/web/src/components/layout/authenticated-layout.tsx
@@ -1,9 +1,14 @@
-import { type CSSProperties, type ReactNode } from "react";
+import { useState, type CSSProperties, type ReactNode } from "react";
 import { useQuery } from "convex/react";
 import { PhoneIcon } from "lucide-react";
 import { useTranslation } from "react-i18next";
+import { toast } from "sonner";
 import type { Id } from "../../../../../convex/_generated/dataModel";
 import { api } from "../../../../../convex/_generated/api";
+import type {
+  BillingInterval,
+  BillingStatus,
+} from "../../../../../packages/shared/src/billing";
 
 import { cn } from "@/lib/utils";
 import { AppSidebar } from "@/components/app-sidebar";
@@ -12,8 +17,14 @@ import { SiteHeader } from "@/components/site-header";
 import { SidebarInset, SidebarProvider } from "@/components/ui/sidebar";
 import { Skeleton } from "@/components/ui/skeleton";
 import { formatPhoneNumberDisplay } from "@/lib/phone";
+import { useObservedAction } from "@/lib/observed-convex";
+import {
+  UpgradePlanDialog,
+  type HostedUpgradePlan,
+} from "@/features/settings/UpgradePlanDialog";
 
 type AuthenticatedLayoutProps = {
+  billingStatus?: BillingStatus;
   businessId?: Id<"businesses">;
   businessName?: string;
   children: ReactNode;
@@ -88,6 +99,7 @@ function AiPhoneNumberPill({
 }
 
 export function AuthenticatedLayout({
+  billingStatus,
   businessId,
   businessName,
   children,
@@ -98,7 +110,47 @@ export function AuthenticatedLayout({
   showUpgradeToPro = false,
   isLoading = false,
 }: AuthenticatedLayoutProps) {
+  const { t } = useTranslation("settings");
   const defaultOpen = getSidebarDefaultOpen();
+  const startCheckout = useObservedAction(api.billing.startCheckout);
+  const [upgradeDialogOpen, setUpgradeDialogOpen] = useState(false);
+  const [upgradeBillingInterval, setUpgradeBillingInterval] =
+    useState<BillingInterval>("annual");
+  const [loading, setLoading] = useState<"checkout" | "portal" | null>(null);
+  const [loadingCheckoutPlan, setLoadingCheckoutPlan] =
+    useState<HostedUpgradePlan | null>(null);
+  const upgradePlans: Array<"starter" | "pro"> =
+    billingStatus?.hasCheckoutAccess === true && billingStatus.plan === "free_cloud"
+      ? billingStatus.availableCheckoutPlans.filter(
+          (plan): plan is "starter" | "pro" =>
+            plan === "starter" || plan === "pro",
+        )
+      : billingStatus?.hasCheckoutAccess === true &&
+          billingStatus.plan === "starter" &&
+          billingStatus.availableCheckoutPlans.includes("pro")
+        ? ["pro"]
+        : [];
+
+  async function handleUpgrade(
+    target: HostedUpgradePlan,
+    billingInterval: BillingInterval,
+  ) {
+    if (!businessId) {
+      return;
+    }
+
+    setLoading("checkout");
+    setLoadingCheckoutPlan(target);
+    try {
+      const result = await startCheckout({ businessId, target, billingInterval });
+      window.location.assign(result.url);
+    } catch {
+      toast.error(t("billing.toast.checkoutFailed"));
+    } finally {
+      setLoading(null);
+      setLoadingCheckoutPlan(null);
+    }
+  }
 
   return (
     <SidebarProvider
@@ -112,12 +164,39 @@ export function AuthenticatedLayout({
       <AppSidebar
         isLoading={isLoading}
         onSignOut={onSignOut}
+        {...(businessId && billingStatus
+          ? {
+              onUpgradeToPro: () => setUpgradeDialogOpen(true),
+            }
+          : {})}
         {...(businessName ? { businessName } : {})}
         {...(operatorAvatar ? { operatorAvatar } : {})}
         {...(operatorEmail ? { operatorEmail } : {})}
         {...(operatorName ? { operatorName } : {})}
         showUpgradeToPro={showUpgradeToPro}
       />
+      {businessId && billingStatus ? (
+        <UpgradePlanDialog
+          availableCheckoutPlans={upgradePlans}
+          billingInterval={upgradeBillingInterval}
+          currentPlan={billingStatus.plan}
+          loading={loading}
+          loadingPlan={loadingCheckoutPlan}
+          onBillingIntervalChange={setUpgradeBillingInterval}
+          onContactEnterprise={() => {
+            window.location.assign(
+              `mailto:hello@lobbystack.ai?subject=${encodeURIComponent(
+                t("billing.upgradeDialog.enterpriseSubject"),
+              )}`,
+            );
+          }}
+          onOpenChange={setUpgradeDialogOpen}
+          onStartCheckout={(target) =>
+            void handleUpgrade(target, upgradeBillingInterval)}
+          open={upgradeDialogOpen}
+          t={t}
+        />
+      ) : null}
       <SidebarInset
         className={cn(
           "@container/content",
